losowanie liczby cakowitej

+++ src/Main.java
@@ -6,7 +6,8 @@ public class Main {
         System.out.println("Hello world!");
         System.out.println("na lekcji programowania");
 
-        double wylosowanaLiczba = Math.random()*100+1;
+        int wylosowanaLiczba = (int) (Math.random()*100+1);
+        //rzutowanie
         System.out.println(wylosowanaLiczba);
     }
 }
